requirements.txt
bson
flask
flask_cors
pymongo
kafka-python
python-decouple
grpcio
google
